winevulkan: Allow vkGetDeviceProcAddr to load instance functions for broken games.
Doom and Wolfenstein II use vkGetDeviceProcAddr to load all their function pointers,
while they are supposed to use vkGetInstanceProcAddr for the instance ones.
This is in violation of the Vulkan spec, but so far the loader allows this, but
Khronos is thinking about fixing the spec and potentially the loader.
https://github.com/KhronosGroup/Vulkan-LoaderAndValidationLayers/issues/2323