winevulkan: Add initial version of vulkan generator script.
The vulkan.py script is what I use to generate the vulkan code.
Khronos uses similar XML files for OpenGL and other standards, but
parsing vk.xml is much more complicated. In particular, because
Vulkan is fond of structures for parameters and it's object orientated
design.

The script parses the XML data into a python object representation,
which I did, because there is so much bookkeeping to do. Each object
typically knows how to parse itself from XML and how to generate C
code for definitions or actual thunks.

Win32 alignment and conversion code will leverage the existing objects
as well. The code is just not ready yet for sharing.